+++ src/shader_recompiler/ir/basic_block.h
@@ -42,7 +42,7 @@ public:
 
     /// Prepends a new instruction to this basic block before the insertion point.
     iterator PrependNewInst(iterator insertion_point, Opcode op,
-                            std::initializer_list<Value> args = {}, u32 flags = 0);
+                            std::initializer_list<Value> args = {}, u64 flags = 0);
 
     /// Adds a new branch to this basic block.
     void AddBranch(Block* block);
